+++ tse_analytics/views/about_dialog_ui.py
@@ -80,8 +80,8 @@ class Ui_AboutDialog(object):
 
         self.buttonBox = QDialogButtonBox(AboutDialog)
         self.buttonBox.setObjectName(u"buttonBox")
-        self.buttonBox.setOrientation(Qt.Horizontal)
-        self.buttonBox.setStandardButtons(QDialogButtonBox.Close)
+        self.buttonBox.setOrientation(Qt.Orientation.Horizontal)
+        self.buttonBox.setStandardButtons(QDialogButtonBox.StandardButton.Close)
 
         self.verticalLayout.addWidget(self.buttonBox)
 
@@ -441,15 +441,13 @@ class Ui_AboutDialog(object):
 "<li style=\" margin-top:12px; margi"
                         "n-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><a href=\"https://pandas.pydata.org/\"><span style=\" text-decoration: underline; color:#0078d4;\">pandas</span></a></li>\n"
 "<li style=\" margin-top:12px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><a href=\"https://pingouin-stats.org/\"><span style=\" text-decoration: underline; color:#0078d4;\">Pingouin</span></a></li>\n"
-"<li style=\" margin-top:12px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><a href=\"https://plotly.com/python/\"><span style=\" text-decoration: underline; color:#0078d4;\">Plotly</span></a></li>\n"
 "<li style=\" margin-top:12px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><a href=\"https://github.com/sbi-rviot/ph_table/\"><span style=\" text-decoration: underline; color:#0078d4;\">pretty_html_table</span></a></li>\n"
-"<li style=\" margin-top:12px; margin-b"
-                        "ottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><a href=\"https://github.com/giampaolo/psutil/\"><span style=\" text-decoration: underline; color:#0078d4;\">psutil</span></a></li>\n"
-"<li style=\" margin-top:12px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><a href=\"https://www.pyqtgraph.org/\"><span style=\" text-decoration: underline; color:#0078d4;\">PyQtGraph</span></a></li>\n"
+"<li style=\" margin-top:12px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><a href=\"https://github.com/giampaolo/psutil/\"><span style=\" text-decoration: underline; color:#0078d4;\">psutil</span></a></li>\n"
+"<li style=\" margin-top:12px"
+                        "; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><a href=\"https://www.pyqtgraph.org/\"><span style=\" text-decoration: underline; color:#0078d4;\">PyQtGraph</span></a></li>\n"
 "<li style=\" margin-top:12px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><a href=\"https://github.com/niklashenning/pyqttoast/\"><span style=\" text-decoration: underline; color:#0078d4;\">PyQt Toast</span></a></li>\n"
 "<li style=\" margin-top:12px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><a href=\"https://seaborn.pydata.org/\"><span style=\" text-decoration: underline; color:#0078d4;\">seaborn</span></a></li>\n"
-"<li style=\" margin-top:12px; m"
-                        "argin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><a href=\"https://github.com/traja-team/traja/\"><span style=\" text-decoration: underline; color:#0078d4;\">Traja</span></a></li></ul></body></html>", None))
+"<li style=\" margin-top:12px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><a href=\"https://github.com/traja-team/traja/\"><span style=\" text-decoration: underline; color:#0078d4;\">Traja</span></a></li></ul></body></html>", None))
         self.tabWidget.setTabText(self.tabWidget.indexOf(self.tabLibraries), QCoreApplication.translate("AboutDialog", u"Third-party libraries", None))
     # retranslateUi