+++ tse_analytics/views/misc/animal_selector.py
@@ -7,7 +7,14 @@ class AnimalSelector(QComboBox):
     def __init__(self, parent=None):
         super().__init__(parent)
 
-    def set_data(self, animals: dict[str, Animal], add_empty_item=True):
+        self.animals: dict[str, Animal] = {}
+
+    def set_data(self, animals: dict[str, Animal], add_empty_item=False):
+        self.animals = animals
         self.clear()
-        items = [""] + list(animals.keys()) if add_empty_item else list(animals.keys())
+        items = [""] + list(animals) if add_empty_item else list(animals)
         self.addItems(items)
+
+    def get_selected_animal(self) -> Animal | None:
+        animal_name = self.currentText()
+        return self.animals[animal_name] if animal_name != "" else None